project_scripts/requirements.txt
aiofiles==23.2.1
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
coloredlogs==15.0.1
contourpy==1.3.2
cuda-bindings==12.9.4
cuda-pathfinder==1.2.2
cycler==0.12.1
einops==0.7.0
exceptiongroup==1.3.1
fastapi==0.135.1
ffmpy==1.0.0
filelock==3.20.0
flatbuffers==25.12.19
fonttools==4.62.1
freetype-py==2.5.1
fsspec==2025.12.0
glcontext==2.5.0
GPUtil==1.4.0
gradio==4.8.0
gradio_client==0.7.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
lazy-loader==0.5
llvmlite==0.46.0
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
mdurl==0.1.2
moderngl==5.10.0
mpmath==1.3.0
narwhals==2.18.0
networkx==3.4.2
numba==0.64.0
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
objaverse==0.1.7
omegaconf==2.3.0
onnxruntime==1.23.2
opencv-python-headless==4.11.0.86
orjson==3.11.7
packaging @ file:///home/task_176104877067765/conda-bld/packaging_1761049113113/work
pandas==2.3.3
Pillow==10.1.0
platformdirs==4.9.4
pooch==1.9.0
protobuf==7.34.0
psutil==7.2.2
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
pyglet==2.1.13
Pygments==2.19.2
PyMatting==1.1.15
PyOpenGL==3.1.0
pyparsing==3.3.2
pyrender==0.1.45
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
rembg==2.0.69
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
rtree==1.4.1
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.15.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
starlette==0.52.1
sympy==1.14.0
tifffile==2025.5.10
tokenizers==0.14.1
tomlkit==0.12.0
torch==2.10.0+cu128
torchaudio==2.10.0+cu128
torchmcubes @ git+https://github.com/tatsy/torchmcubes.git@3381600ddc3d2e4d74222f8495866be5fafbace4
torchvision==0.25.0+cu128
tqdm==4.67.3
transformers==4.35.0
trimesh==4.0.5
triton==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
websockets==11.0.3
xatlas==0.0.9
